+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 * Add your own contributions to the next release on the line below this, please include your name too. Please don't set a new version if you are the first to make the section for `master`.
 
+## 0.1.1
+
+- Fix install paths for libDanger when using homebrew - orta
+
+## 0.1.0
+
+- First release via homebrew - eneko
+
+
 ## 0.0.2
 
 * Supports a Dangerfile in both: "/Dangerfile.swift", "/danger/Dangerfile.swift" or "/Danger/Dangerfile.swift" to handle SPM rules on Swift files in the root - orta